+++ product_code/certificate_service/impl/src/main/scala/de/upb/cs/uc4/certificate/impl/CertificateSerializerRegistry.scala
@@ -1,7 +1,9 @@
 package de.upb.cs.uc4.certificate.impl
 
 import com.lightbend.lagom.scaladsl.playjson.JsonSerializer
+import de.upb.cs.uc4.certificate.impl.actor.CertificateState
 import de.upb.cs.uc4.certificate.impl.events.{ OnCertficateAndKeySet, OnCertificateUserDelete, OnRegisterUser }
+import de.upb.cs.uc4.certificate.model.EncryptedPrivateKey
 import de.upb.cs.uc4.shared.server.SharedSerializerRegistry
 
 import scala.collection.immutable.Seq
@@ -18,6 +20,8 @@ object CertificateSerializerRegistry extends SharedSerializerRegistry {
   override def customSerializers: Seq[JsonSerializer[_]] = Seq(
     JsonSerializer[OnRegisterUser],
     JsonSerializer[OnCertficateAndKeySet],
-    JsonSerializer[OnCertificateUserDelete]
+    JsonSerializer[OnCertificateUserDelete],
+    JsonSerializer[CertificateState],
+    JsonSerializer[EncryptedPrivateKey]
   )
 }