+++ product_code/course_service/impl/src/main/scala/de/upb/cs/uc4/course/impl/CourseServiceImpl.scala
@@ -18,7 +18,7 @@ import de.upb.cs.uc4.course.model.Course
 import de.upb.cs.uc4.examreg.api.ExamregService
 import de.upb.cs.uc4.shared.client.exceptions._
 import de.upb.cs.uc4.shared.server.ServiceCallFactory._
-import de.upb.cs.uc4.shared.server.messages.{ Accepted, Confirmation, Rejected, RejectedWithError }
+import de.upb.cs.uc4.shared.server.messages.{ Accepted, Confirmation, Rejected }
 import de.upb.cs.uc4.user.api.UserService
 
 import scala.concurrent.duration._
@@ -143,7 +143,7 @@ class CourseServiceImpl(
               .map {
                 case Accepted => // Creation Successful
                   (ResponseHeader(201, MessageProtocol.empty, List(("Location", s"$pathPrefix/courses/${courseToAdd.courseId}"))), courseToAdd)
-                case RejectedWithError(code, reason) =>
+                case Rejected(code, reason) =>
                   throw UC4Exception(code, reason)
               }
           }
@@ -167,8 +167,8 @@ class CourseServiceImpl(
                   .map {
                     case Accepted => // OK
                       (ResponseHeader(200, MessageProtocol.empty, List()), Done)
-                    case Rejected(reason) => // Not Found
-                      throw UC4Exception.InternalServerError("Course Deletion Error", reason)
+                    case Rejected(code, reason) => // Not Found
+                      throw UC4Exception(code, reason)
                   }
               }
             case None =>
@@ -257,7 +257,7 @@ class CourseServiceImpl(
                   .map {
                     case Accepted => // Update Successful
                       (ResponseHeader(200, MessageProtocol.empty, List()), Done)
-                    case RejectedWithError(code, reason) =>
+                    case Rejected(code, reason) =>
                       throw UC4Exception(code, reason)
                   }
               }